relocate functions to progressive files
