+++ internal/controller/pipelinerollout/pipelinerollout_progressive.go
@@ -9,7 +9,9 @@ import (
 
 	numaflowv1 "github.com/numaproj/numaflow/pkg/apis/numaflow/v1alpha1"
 	"github.com/numaproj/numaplane/internal/common"
+	"github.com/numaproj/numaplane/internal/controller/common/numaflowtypes"
 	"github.com/numaproj/numaplane/internal/controller/progressive"
+	"github.com/numaproj/numaplane/internal/util"
 	"github.com/numaproj/numaplane/internal/util/kubernetes"
 	"github.com/numaproj/numaplane/internal/util/logger"
 	apiv1 "github.com/numaproj/numaplane/pkg/apis/numaplane/v1alpha1"
@@ -139,6 +141,39 @@ func (r *PipelineRolloutReconciler) AssessUpgradingChild(ctx context.Context, ro
 	}
 
 	return progressive.AssessUpgradingPipelineType(ctx, pipelineRollout.GetAnalysisStatus(), existingUpgradingChildDef, verifyReplicasFunc)
+}
+
+// UpgradingChildNeedsUpdating() tests for essential equality, with any fields that Numaplane manipulates eliminated from the comparison
+// This implements a function of the progressiveController interface, used to determine if a previously Upgrading Pipeline
+// should be replaced with a new one.
+// What should a user be able to update to cause this?: Ideally, they should be able to change any field if they need to and not just those that are
+// configured as "progressive", in the off chance that changing one of those fixes a problem.
+// However, we need to exclude any field that Numaplane itself changes or it will confuse things.
+// TODO: move to pipelinerollout_progressive.go
+func (r *PipelineRolloutReconciler) UpgradingChildNeedsUpdating(ctx context.Context, from, to *unstructured.Unstructured) (bool, error) {
+	numaLogger := logger.FromContext(ctx)
+	fromCopy := from.DeepCopy()
+	toCopy := to.DeepCopy()
+
+	err := numaflowtypes.PipelineWithoutScaleMinMax(fromCopy)
+	if err != nil {
+		return false, err
+	}
+	err = numaflowtypes.PipelineWithoutScaleMinMax(toCopy)
+	if err != nil {
+		return false, err
+	}
+
+	specsEqual := util.CompareStructNumTypeAgnostic(fromCopy.Object["spec"], toCopy.Object["spec"])
+	numaLogger.Debugf("specsEqual: %t, from=%v, to=%v\n",
+		specsEqual, fromCopy.Object["spec"], toCopy.Object["spec"])
+	// compare Labels and Annotations, excluding any that Numaplane itself applies
+	labelsEqual := util.CompareMapsWithExceptions(from.GetLabels(), to.GetLabels(), common.KeyNumaplanePrefix)
+	numaLogger.Debugf("labelsEqual (excluding Numaplane labels): %t, from Labels=%v, to Labels=%v", labelsEqual, from.GetLabels(), to.GetLabels())
+	annotationsEqual := util.CompareMapsWithExceptions(from.GetAnnotations(), to.GetAnnotations(), common.KeyNumaplanePrefix)
+	numaLogger.Debugf("annotationsEqual (excluding Numaplane annotations): %t, from Annotations=%v, to Annotations=%v", annotationsEqual, from.GetAnnotations(), to.GetAnnotations())
+
+	return !specsEqual || !labelsEqual || !annotationsEqual, nil
 }
 
 /*
